Added force-directed graph to collab search.
A new D3.js visualization has been added. This is the first visual outside of
nvd3. It is currently a little rough arround the edges but should server as a
strong prototype and example for many more visualizations.

Currently the JS is included on every 'page'
Debug statements remain as comments. [will remov]