Created authorArrayFDG function.
This function can produce a force-direct graph with any array of author lists.
By making it a separate function I can enable different searches that produce
the same type of visualization. This sort of separation is the goal.
This function could likely be extended to take other types of arrays of lists.